// src/shared/common/types.ts
export interface SymbolItem {
    name: string;
    detail: string;
    kind: number; // vscode.SymbolKind
    range: any; // vscode.Range
    selectionRange: any; // vscode.Range
    children: SymbolItem[];
    uri?: string; // For workspace symbols
    containerName?: string;
    autoExpand?: boolean;
    isDeepSearch?: boolean;
    path?: string;
}